Move `file_watcher` to `revproxy/watcher` package for better modularity.
